(refactor) rename PostTags to BlogPostTags
